+++ app/src/main/java/com/example/text_it/Activity/LoginUser.kt
@@ -12,12 +12,6 @@ class LoginUser : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_login_user)
 
-        if (checkSelfPermission(android.Manifest.permission.CAMERA) != android.content.pm.PackageManager.PERMISSION_GRANTED) {
-            requestPermissions(arrayOf(
-                android.Manifest.permission.CAMERA
-            ), 1)
-        }
-
         val loginBut: Button = findViewById(R.id.buttonLogin)
         val regBut: Button = findViewById(R.id.buttonRegister)
 
@@ -31,18 +25,6 @@ class LoginUser : AppCompatActivity() {
 
         loginBut.setOnClickListener {
 //            implement the logic for logging in the user using firebase
-
-        }
-    }
-    override fun onRequestPermissionsResult(
-        requestCode: Int,
-        permissions: Array<out String>,
-        grantResults: IntArray
-    ) {
-        super.onRequestPermissionsResult(requestCode, permissions, grantResults)
-        if(grantResults[0] != android.content.pm.PackageManager.PERMISSION_GRANTED){
-            Toast.makeText(this, "Permissions not granted", Toast.LENGTH_SHORT).show()
-            finish()
         }
     }
 }
