+++ app/src/main/java/com/example/text_it/Activity/SplashActivity.kt
@@ -18,14 +18,14 @@ class SplashActivity : AppCompatActivity() {
 
         iv_logo.alpha = 0f
         iv_logo.animate().setDuration(3000).alpha(1f).withEndAction {
-            val i = Intent(this, LoginUser::class.java)
+            val i = Intent(this, LandingUser::class.java)
             startActivity(i)
             overridePendingTransition(android.R.anim.fade_in, android.R.anim.fade_out)
             finish()
         }
 
         Handler().postDelayed({
-            val i = Intent(this, LoginUser::class.java)
+            val i = Intent(this, LandingUser::class.java)
             startActivity(i)
             overridePendingTransition(android.R.anim.fade_in, android.R.anim.fade_out)
             finish()
