+++ app/src/main/res/drawable/stylized_circle.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="oval">
+<solid android:color="#33ffffff"/>
+    <size android:width="48dp" android:height="48dp"/>
+</shape>
