+++ app/src/main/res/drawable/stylized_gradient_button.xml
@@ -3,7 +3,7 @@
     <item >
         <shape android:shape="rectangle"  >
             <corners android:radius="16dp" />
-            <solid android:color="#33ffffff"/>
+            <gradient android:startColor="#000001" android:endColor="#4055A3" />
         </shape>
     </item>
 </selector>
